// src/utils/models/notifications/cartelera.ts
export class Billboard{
    title: string;
    description: string;
    images: any;
    state: string;
    start_date: string;
    end_date: string;
    author: string;
    link: string;
    btn_text: string;
    id_establishment: string;
    _id: string;
    updated_at: string;
    is_important: string;
    created_at: any;
}